+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - **Breaking:** `Device::name`, `Device::output_name`, `Seat::physical_name` and `Seat::logical_name` now return Cow strings due to lossy conversion into UTF8
+- Added support for libinput accel config
 
 ## 0.9.1